requirements.txt
# Jarves Dependencies

# Core
pyyaml>=6.0
psutil>=5.9.0

# AI/LLM
requests>=2.28.0

# Voice (optional)
SpeechRecognition>=3.10.0
pyttsx3>=2.90
PyAudio>=0.2.13

# Web interface (optional)
flask>=2.3.0

# Security
pycryptodome>=3.18.0
